# public/doc/swagger.yaml
swagger: '2.0'
info:
  description: HFS API ドキュメント
  version: 0.0.1
  title: HFS Platform REST APIs
host: localhost:3000
basePath: /api
tags:
  - name: hfs
    description: HFS APIs
schemes:
  - http
  - https
paths:
  /file/{id}:
    get:
      tags:
        - hfs
      summary: 指定した１つのデータを取り出す
      description: \{id\} で指定した id のメタデータを一件取りだす
      parameters:
        - name: id
          type: string
          in: path
          description: id
          required: true
        - name: binary
          type: string
          in: query
          description: 指定した id のバイナリデータを Content-Type ごと取り出す
        - name: download
          type: string
          in: query
          description: 指定した id のデータをダウンロードする
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
    delete:
      tags:
        - hfs
      summary: 指定した１つのデータを削除する
      description: \{id\} で指定した id のデータを一件削除する
      parameters:
        - name: id
          type: string
          in: path
          description: id
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /file_ids:
    get:
      tags:
        - hfs
      summary: データを全件取り出す
      description: 全データの id を取りだす
      parameters:
        - name: limit
          type: number
          in: query
          description: 取得するデータ数
        - name: offset
          type: number
          in: query
          description: 取得するデータのオフセット
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /file:
    post:
      tags:
        - hfs
      summary: データを作成する
      description: データを作成する（データのハッシュ値が _id 値として記録される）
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: file
          type: file
          description: アップロードするファイル
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /exist:
    post:
      tags:
        - hfs
      summary: アップロードするデータが存在しているかを確認する
      description: アップロードするデータが存在しているかを確認する
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: file
          type: file
          description: アップロードするファイル
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /validate/{id}:
    post:
      tags:
        - hfs
      summary: 格納データが改ざんされていないことを確認
      description: ハッシュ値と比較して格納データが改ざんされていないことを確認
      parameters:
        - name: id
          type: string
          in: path
          description: id
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー